+++ iris_grib/tests/unit/save_rules/test_grid_definition_template_12.py
@@ -105,24 +105,6 @@ class Test(tests.IrisGribTest, GdtTestMixin):
         self._check_key("Di", 200)
         self._check_key("Dj", 500)
 
-    def test__negative_grid_points_eccodes_broken(self):
-        self.mock_eccodes.CodesInternalError = FakeGribError
-
-        # Force the test to run the signed int --> unsigned int workaround.
-        def set(grib, key, value):
-            if key in ["X1", "X2", "Y1", "Y2"] and value < 0:
-                raise self.mock_eccodes.CodesInternalError()
-            grib.keys[key] = value
-
-        self.mock_eccodes.codes_set = set
-
-        test_cube = self._make_test_cube(x_points=[-1, 1, 3, 5, 7], y_points=[-4, 9])
-        grid_definition_template_12(test_cube, self.mock_grib)
-        self._check_key("X1", 0x80000064)
-        self._check_key("X2", 700)
-        self._check_key("Y1", 0x80000190)
-        self._check_key("Y2", 900)
-
     def test__negative_grid_points_eccodes_fixed(self):
         test_cube = self._make_test_cube(x_points=[-1, 1, 3, 5, 7], y_points=[-4, 9])
         grid_definition_template_12(test_cube, self.mock_grib)
@@ -138,25 +120,7 @@ class Test(tests.IrisGribTest, GdtTestMixin):
         self._check_key("XR", 40000000.0)
         self._check_key("YR", -10000000.0)
 
-    def test__scale_factor_eccodes_broken(self):
-        # GRIBAPI expects a signed int for scaleFactorAtReferencePoint
-        # but it should accept a float, so work around this.
-        # See https://software.ecmwf.int/issues/browse/SUP-1100
-
-        def get_native_type(grib, key):
-            assert key == "scaleFactorAtReferencePoint"
-            return int
-
-        self.mock_eccodes.codes_get_native_type = get_native_type
-        grid_definition_template_12(self.test_cube, self.mock_grib)
-        self._check_key("scaleFactorAtReferencePoint", 1065346526)
-
-    def test__scale_factor_eccodes_fixed(self):
-        def get_native_type(grib, key):
-            assert key == "scaleFactorAtReferencePoint"
-            return float
-
-        self.mock_eccodes.codes_get_native_type = get_native_type
+    def test__scale_factor(self):
         grid_definition_template_12(self.test_cube, self.mock_grib)
         self._check_key("scaleFactorAtReferencePoint", 0.9996012717)